requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
azure-storage-blob==1.4.0
azure-storage-common==1.4.2
azure-functions
azure-batch==6.0.0 
azure-keyvault==4.1.0
azure-identity==1.7.0